implementing disk operations
