+++ Cargo.toml
@@ -32,6 +32,7 @@ mockall = "0.11"
 semver = "1.0.13"
 nix = "0.25.0"
 uuid = { version = "1.1", features = ["v4"] }
+ioctls = "0.6"
 
 [dev-dependencies]
 hex = "0.4.3"
@@ -40,4 +41,3 @@ hexdump = "0.1.1"
 simple_logger = "2.3"
 rand = "0.8.5"
 tempfile = "3"
-
